[FIX] approvals_purchase: use vendor's currency for rfq
Issue:
When creating an RFQ from an approval, the created purchase order
does not use the currency set on the vendor of the product. Rather,
it uses the currency of the company, with the value converted
based on the vendor's currency to get the price. This is not
consistent with other ways we create RFQs, which all respect
the vendor currency.

Solution:
Pass the vendor's currency into the values sent when creating
the purchase order. In the case of modifying an existing purchase
order, only modify purchase orders matching vendor's currency.

opw-4549937

closes odoo/enterprise#98205

X-original-commit: 7e3376256a0b5042abb67b16304c637e0d75d008
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Austin Paskerian (paau) <paau@odoo.com>