[FIX] web_studio: make the border visible in app creator
Before this commit :
when creating an app using the studio, none of the borders were visible,
including the checkboxes in the app creator, because the background color of the
app creator and the border color were the same.

After this commit:
the border color for the app creator has been changed to make it visible.

task-5138669

closes odoo/enterprise#97454

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>