[FIX] ai: prevent error when clicking on AI icon
Currently, an error occurs when the user clicks on the AI icon at the top.

Steps to Reproduce:
 - Install the website_blog module.
 - Go to Website > Configuration > Blog > Blogs.
 - Open any blog and click the AI icon at the top.

AttributeError: 'blog.blog' object has no attribute 'activity_ids'

This error occurs after [this commit]. When the user clicks the AI icon,
it attempts to create an AI channel and fetch planned activities if any
exist. However, the blog.blog model does not have the activity_ids
attribute because it does not inherit from mail.activity.mixin. As a
result, the error is raised [1].

This commit ensures that the system only checks for activities if the
record has the activity_ids attribute; otherwise, it returns empty
data, indicating no activities found.

[1]: https://github.com/odoo/enterprise/blob/64fa9e46f6195bd4312130cce33b94a129584523/ai/models/mail_thread.py#L12

[this commit]: https://github.com/odoo/enterprise/commit/7333dc779c75479c708edb7ecd1c0a0d4d2ae619

sentry-6942245851

closes odoo/enterprise#97136

Signed-off-by: Panagiotis Kyriakou (paky) <paky@odoo.com>