[FIX] {account,l10n_ro}_saft: TaxInformation with reverse charge
When a reverse charge tax is used, we export a `TaxInformation` with a
zero amount instead of the real amount of the tax.
The tax authority requires to show the tax to pay and doesn't care about
the tax to receive...

In order to fix this, we only sum the tax details of lines having
positive repartition lines.

opw-5125678

closes odoo/enterprise#97435

X-original-commit: 9612a8ba00f2bc86aa6b33d64e5fdcfe5f841bc5
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: William André (wan) <wan@odoo.com>