[FIX] pos_iot: avoid permission error for non-POS user
Before this commit, if an IoT user/admin attempted to access any IoT
view while `pos_iot` was installed, they would get a permission error if
they were not also a POS user.

After this commit, we check if the user is in a POS group before showing
any POS related fields. We also check the group when autoconfiguring a
new IoT box with a POS, this feature will simply be skipped if the user
cannot access POS.

closes odoo/enterprise#97225

X-original-commit: f8090a1002e3345670709dd921cbe861dcee8937
Signed-off-by: Louis Travaux (lotr) <lotr@odoo.com>
Signed-off-by: Max Whale (mawh) <mawh@odoo.com>