[FIX] l10n_*: Correct state of returns of a tax report.
It's possible for different tax reports to exist in the same country, each with its own return type. However, this is an issue when they aren't configured to generate closing entries with distinct contents, since each of them will then repeat the content of the other.

To fix this, when it's legitimate to make multiple closing entries, we make sure the closing entry only considers the tax move lines linked to the report's tax tags.

If multiple reports must generate the same closing entry (typically because one of them is a simplified version of the other), we disable the generation of the return for one of them.

Some reports also didn't need to generate a closing entry and were switched to 'generic_state_review_submit'.

task-5123564

closes odoo/enterprise#99182

X-original-commit: dcfef2cc462631f376a596a7c85ae483826835ad
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Léo Leclerc (leol) <leol@odoo.com>