[IMP] l10n_pe_edi: support withholding tax in ubl
Some customers in Peru are considered "agentes de retención", and as
such have special considerations that are necessary. We must perform
specific actions when working with these customers, when selling to them
we must:
- recieve a specific document from the customer (A followup task will
  add this support for companies to generate this document)
- Include a retention reference in our XML.

This is represented as a specific `AllowanceCharge` node that is the sum
of all of the Withholding taxes on the lines with the specific code of
62.

We know an invoice is to a withholding customer because all lines must
have a tax that is a part of the withholding tax group.

This withholding tax is not included in the monetarytotals, tax totals,
or invoice line values. And there can be only one AllowanceCharge node
per invoice, not one per line.

task-4989463

closes odoo/enterprise#97071

X-original-commit: fce157b2242853b10a54f347f1c8078b6bff888e
Related: odoo/odoo#231351
Signed-off-by: Joren Van Onder (jov) <jov@odoo.com>
Signed-off-by: Andrew Gavgavian (andg) <andg@odoo.com>