[FIX] l10n_in_hr_payroll: fixes runbot build error
The issue occurs because the _get_whitelist_fields_from_template() method
includes the overtime_from_attendance field, introduced in this #92093.
This field comes from the hr_work_entry_attendance module, which is not listed
as a dependency in all payroll localisations. As a result, the field cannot be
found during test execution.

Solution:
Remove the overtime_from_attendance field from all
_get_whitelist_fields_from_template overrides in the localisation modules,
and instead include this field by overriding the function in
the hr_work_entry_attendance module.

build error-243483
task-5169211

closes odoo/enterprise#97201

Signed-off-by: Jurgen Gjini (jugj) <jugj@odoo.com>