[FIX] base: wrap the user name in their signature in a `div`
Adapt signatures for Laurie Poiret and Maggie Davidson so that they are wrapped
in a `<div>` instead of a `<p>`, because lxml `document_fromstring` would
wrap text content in a `p` during record creation. Ideally all user data
records should do that for their signature field. This commit only updates the
main ones.

`p` in a signature is not desired because it has a native margin-bottom, and the
user has to change it to a `div` in order not to have it. Making `div` the
default container for signatures simplifies that tedious process.

task-5149570

closes odoo/enterprise#97229

Related: odoo/odoo#230275
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>