[FIX] delivery_starshipit: missing city fields
Current behaviour:
When using Starshipit, city is not included
in the payload sent to the API. This causes
an error for some delivery services.

Steps to reproduce:
1. Install delivery_starshipit
2. Create starshipit Shipping Method (using Starshipit API key)
3. Set up only "NZ Post Domestic" in Starshipit portal.
4. Click on "Select a service linked to your starshipit account"
5. Error: There are no shipping services available

Cause of the issue:
Some delivery services (like NZ Post Domestic) require a city/suburb
field. In Australia/New Zealand suburb is used in the city field.

opw-5155401

closes odoo/enterprise#97588

X-original-commit: f4b430c1cc4523572b8a933d8563169bfffabc0f
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>
Signed-off-by: Jonathan Andika (jand) <jand@odoo.com>