[IMP] account_fiscal_categories: enhance list view of account_account
With this commit, we add a new field `current_rate` in `account.account`. This
`current_rate` is a non-stored compute field, used to compute the current
applicable rate on the given account.

This `current_rate` and `fiscal_category_id` has been introduced in the list
view of `account.account`.

task-5163392

Part-of: odoo/enterprise#97324
Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>