[FIX] project_sale_subscription: fix subscription section in dashboard
Before this commit, when subscriptions were linked to the analytic
account of a project, the sale order items appears in an unwanted
section when the section is unfolded. Meanwhile, when the subscription
section is unfolded the title of the subscriptions items are not
correctly displayed.

The first issue is due to the fact that we did not correctly exclude the
subscriptions items from the domain.
The second issue is due to the fact that we fetch the field 'name' from
the subscription search instead of the field 'display_name'

task-5159781

closes odoo/enterprise#101614

X-original-commit: 374d115252f7183fbec44b6b9fb6d81c735906a7
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>