[FIX] mrp_workorder_hr_account: fix cost for done wo
This commit fixes the problem of adding the hourly cost of the
work center when marking an operation as done from the status widget.

To reproduce the bug:
1- Create a work center with an hourly cost > 0.
2- Create an MO with 1 operation in that work center, expected time > 0.
3- Create and set a project on the MO.
4- Make sure that project has an analytic account.
5- Mark the operation as done from the status widget.
(click on it and choose done, don't use the start button)
6- Go to the analytic account of the project and check the gross margin.
= No cost of the workorder was added.

Now, this commit takes into account the duration of the WO
first when marking it as done directly from the status widget.

Task-5156286

closes odoo/enterprise#97253

Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>