[IMP] account_asset: set depreciation schedule default opening previous year
The depreciation schedule report is usually is consulted to be attached to the
BNB and/or tax return of a completed financial year. For an ongoing financial
year, it has very little value.

Therefore default opening of the depreciation schedule report is changed from
previous month(default value) to previous year.

task-5167082

closes odoo/enterprise#97327

Signed-off-by: William André (wan) <wan@odoo.com>