[FIX] appointment: handle invalid booking URLs gracefully
This error occurs when a user provides an invalid URL
(e.g., one containing characters like “:”, “..”, etc.) in the link field.

Steps to reproduce:
- Install 'appointment' module
- Appointment > Shared Links > New
- Fill 'Appointment Types' > Link: 'testurl:' (provide wrong link in Extra link)

Traceback:
ValueError: Extra URL must use same scheme and host as base, and begin with base path

At [1], this error occurs because during computation, the method calls
'urljoin'. If the URL contains a colon (e.g., ‘:’), it is compared with the
base URL, and when they differ, a 'ValueError' is intentionally raised at [2].
This leads to the observed error.

This commit resolves the issue by catching the error when an invalid URL is
provided.

[1]: https://github.com/odoo/enterprise/blob/fc931c81ca7b2ebad220cf057e15583e046810b9/appointment/models/appointment_invite.py#L280

[2]: https://github.com/odoo/odoo/blob/9900375bc0bac2754150fd8cd6a1e45ecad8da83/odoo/tools/urls.py#L55-L58

sentry-6941171133

closes odoo/enterprise#97353

Signed-off-by: Jérémy Hennecart (jeh) <jeh@odoo.com>