[FIX] social_youtube: remove jquery for youtube upload field
Issue: when trying to upload a video to youtube, we receive an error as
"ReferenceError: $ is not defined" meaning we're not able to reach the
jQuery identifier.

So following the trend we're moving away from jQuery, this commit
removes the usage of jQuery in the youtube upload field.

- Replaced $.ajax() with fetch() in _openUploadSession,
_updateProcessingInfo, and _onClearClick methods

- Replace $.ajax() with XMLHttpRequest in _uploadFile
to maintain upload progress tracking

opw-5142810

closes odoo/enterprise#97364

Signed-off-by: Julien Banken (jbn) <jbn@odoo.com>