[FIX] delivery_iot: auto print multiple documents
The `button_validate` method called clicking "Validate" returns a list
of client actions to call. After these clients actions are executed,
the page reloads.

This reload makes our broadcasted action not to be caught by the client
if there is only one connected.

Anyway, this flow was overcomplicated and has been simplified overriding the
method returning the client actions, adding the "shipping labels" and "shipping documents"
to it.

closes odoo/enterprise#97761

X-original-commit: b8d0ee2c43a67a8fbafc8a4c7ea54c16f203753f
Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>
Signed-off-by: Louis Travaux (lotr) <lotr@odoo.com>