[FIX] helpdesk: create ticket from mail if partner in other company
**Steps to reproduce**
- Have a helpdesk team for company A, with an email alias set up.
- Have a `res.partner` belonging to company B with an email. To reproduce,
there should be no contact with this email in company A or with no company.
- Send a message with the email of the `res.partner` to the alias
of the helpdesk team.
Issue: the ticket is not created.

**Cause**
The `_check_partner_id_has_the_same_company` constraint fails, preventing
the ticket from being created.

**Solution**
We only use `author_id` if it respects the constraint.
https://github.com/odoo/enterprise/blob/a1fe558c8d6e27b435d350c23e19b93bbe611eb7/helpdesk/models/helpdesk_ticket.py#L755
This values comes from `_mail_find_partner_from_emails` in the mail
module, which will prioritise a contact from the same company, but
will not prevent one from another company being used.
The existing code in `create` will call `_find_or_create_partner`
to find/create one based on the name and mail.
`_mail_find_partner_from_emails` is used instead of `find_or_create`
to be able to provide an additional domain to ensure company consistency.

opw-5154449

closes odoo/enterprise#98133

X-original-commit: ec803c3ab6fba191aaac04ab59204ae79460d50a
Signed-off-by: Tanguy Quéguineur (taqu) <taqu@odoo.com>
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>