[FIX] delivery_fedex_rest: adapt test
Once the put in pack is done, the order of the move_lines changes (due
to their `_order` clause). This meant that the same move line was again
updated right after.

This didn't cause issue before because when one line was picked and the
other not, when doing another put_in_pack, it would have selected the
un-picked one before selecting the picked one and do a pack in pack.

Since this is no longer the case, we need to make sure we have have the
right lines.

Task-5116567

closes odoo/enterprise#95648

Related: odoo/odoo#228837
Signed-off-by: William Henrotin (whe) <whe@odoo.com>