[FIX] ai: fix exception handling when using generate_response
Before this commit, whenever we called `/ai/generate_response` from the fron-end
we were catching any potential exceptions and calling the `ai/post_error_message`.
The `post_error_message` function called from that controller endpoint would
take the exception message, and ask the AI to explain the error to the user without
the use of technical terms.

The problem with this approach is two-fold.
Firstly, in the `post_error_message` method, we use the `generate_response` method
in order for the AI to beauty-fy the error message. But, the `post_error_message`
method is called when catching exceptions of the `generate_response` method. Thus,
the exception is caught, a nice message is posted in the chat, but then a traceback
is shown regardless.
Secondly, beautifying the error messages makes it more difficult for end users to
understand what could be going wrong and making it also more difficult for our
support to help them out.

In this commit, we removed the `post_error_message` flow. We move the try-except
to the `_generate_response_for_channel` method and if the user is an internal user
we let the exception bubble up. If not (for website users on livechat), a generic
message will be posted on the chat.

task-5177169

closes odoo/enterprise#97516

Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>