[FIX] account_accountant: hide the option "Enable Auto Transfers"
Steps to reproduce:
- Install "account_accountant"
- Go in the settings
- Search for "Enable Auto Transfers", tick the option
- Save
- The option is not ticked

Cause:
This option is supposed to install the module for Account Transfers. The option field is called `module_account_auto_transfer`. It's supposed to install the module `account_auto_transfer`.

But this module was [refactored in 19.0](https://github.com/odoo/enterprise/commit/77c248981c9e1ccbb5d4c88a0915a231f040df84) and renamed `account_transfer`. So the settings can't find the module to install, nothing happen, and the option stays the same.

Solution:
Hide the option in 19.0 to respect the stable: we cannot rename a stored field. This is not a very important loss the users can download the module `account_transfer` directly from the Apps application.

Give the variable the correct name in master (`module_account_transfer`).

opw-5149342

closes odoo/enterprise#97513

Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>