[FIX] stock_barcode_picking_batch: ignore scans during batch creation
Issue:

Scans processed during picking batch creation in the barcode app are
interpreted as scans related to the BarcodePickingBatchModel edition
(that is product, lot, package,... scans) which does not make sense and
raises multiple errors.

Steps to reproduce:

- In the settings enable "Batch Transfers"
- Inventory/Operations/Transfers/Batch Transfers
- Create a new batch and do not confirm it
- Go to the barcode app > Batch Transfers > Clear filters
- Select your draft batch
- Scan anything
> Traceback

Other issues:
The same flow with a cancelled batch triggers the same tracebacks and,
with a done batch triggers the message: "This picking is already done".

Cause of the issue:

Since `this.state.view === "barcodeLines"`, scans are processed as if
we were processing an existing `BarcodePickingBatchModel` with lines:
https://github.com/odoo/enterprise/blob/ce7460b27028d4133c4689afd2e8f303268494a0/stock_barcode/static/src/components/main.js#L204-L209
In particular, if the scan corresponds to a product, and error will be
raised because the scan can not process a `createNewLine` for the
scanned product in the current view and if the scan does not
correspond to anything valid, it will raise an error because
`this.picking` is undefined and `this.picking.use_existing_lots` raises
an error:
https://github.com/odoo/enterprise/blob/ce7460b27028d4133c4689afd2e8f303268494a0/stock_barcode/static/src/models/barcode_model.js#L1115
https://github.com/odoo/enterprise/blob/ce7460b27028d4133c4689afd2e8f303268494a0/stock_barcode_picking_batch/static/src/models/barcode_picking_batch_model.js#L177-L179

opw-5162164

closes odoo/enterprise#98632

X-original-commit: 70fc0d593aca4b3016257e89da2a3033d6aa40c5
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>