[FIX] sale_shopee: Fix unhandled delivery status
When the delivery order is delivered but not yet confirmed by the buyer, an error state is set in the delivery order.

'TO_CONFIRM_RECEIVE' should be included in the delivery status. It refers that the package is delivered but not confirmed by the buyer. It should not be regarded as an error state in Odoo.

task_id: 4965896

closes odoo/enterprise#97781

X-original-commit: 16c662766e3269a97ef7e105a3ba82225a073568
Signed-off-by: Steven Tang (stta) <stta@odoo.com>