[FIX] esg: correctly render graphs in ESG dashboard in dark mode
Before this commit, the color chosen for the graphs in the ESG dashboard
does not take into account the dark/light mode and so they does not have
the right color when the user is in dark mode.

This commit makes sure the graphs are well rendered in dark mode.

task-5138512

Part-of: odoo/enterprise#97646
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>