[FIX] account_accountant: deferred with receipt type
## Steps to reproduce

### Bug 1
1. Create a bill receipt
2. Add deferred dates
3. They're reset to False

### Bug 2
1. Create a bill
2. Set the deferred dates
3. Switch to receipt type
4. Confirm
5. The account of the generated deferral entries is incorrect

## Fix

While checking the document type with the `is_purchase_document` and
`is_sale_document` helper methods, the receipts were ignored as
this is the default value.

In bug 1, this means that the method `_has_deferred_compatible_account`
method would always return `False` when using the receipt type,
therefore reseting the deferred dates.

In bug 2, this means that when generating the deferrals entries, the
`deferred_type` would always be `revenue` in case of a receipt because
of the ternary operator.

For both bugs, we can simply set `include_receipts` to `True` to take
these into account while veryfing/setting the account.

opw-5129561

closes odoo/enterprise#97877

X-original-commit: 42d2d4e78b559799d710fc12107ad6cd1be90182
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>