[FIX] hr_payroll: test is broken in new chrome
Since either 140 or 141[^0] Chrome uses compact serialization rules
for text-decoration, as a result of w3c/csswg-drafts#12486. This was
merged [on August 12th][CL-6824265] and Chrome 139 was cut on August 5
so couldn't have had this change.

The new serialization rule means values which can be ommitted from a
shorthand (default or fallback) should be, so the default
`text-decoration-line` and a `text-decoration-color` matching the
`color` won't be serialized in `text-decoration` anymore.

Instead of matching the shorthand, match the actual stored properties.

[^0]: 140 was not deployed on chrome next so it may or may not have
  been broken then

[CL-6824265]:
  https://chromium-review.googlesource.com/c/chromium/src/+/6824265

closes odoo/enterprise#97667

Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>