[FIX] account_reports: Use right options in test_generate_all_export_files
The aim of this commit is making sure that we are using
the right report options in the test_generate_all_export_files.
Indeed, before this commit, we were using the options generated
before we entered the first dispatch_report_action call.
In this, we could call the export_file function which is returning
a "ir_actions_account_report_download" action. This action adds
a key in the options dict "export_mode": "file". This is necessary
to have the lines' codes in the line data.

no task id

Related: odoo/odoo#229876
Related: odoo/odoo#232543
Related: odoo/odoo#232757
X-original-commit: 044f83b0889340ebd010ca3569d2af062aa75ebc
Part-of: odoo/enterprise#97909
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>