[FIX] account_accountant_batch_payment: skip tests without entries
We have tests for payment wihtout journal entries.
But these payments can only be generated if accountant in installed.
It makes these tests fail if the module is not installed.

To avoid making a bridge module only to test these, we add a class
that will be skipped if the module is not installed.

runbot-232728

closes odoo/enterprise#97764

Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>