[FIX] l10n_be_reports: fields access rights fixed
This commit fixes the access right problem for `l10n_be_region_id`
and `l10n_be_company_type_id` when `accountant` is not installed.
When `accountant` is not installed and users have `l10n_be_reports`
installed, an access error was raised when these fields were accessed.

A  group `account.group_account_user` is added to these fields such
that they become invisible when when `accountant` is not installed.

task-5156687

closes odoo/enterprise#97759

Signed-off-by: William André (wan) <wan@odoo.com>