[FIX] project_timesheet_forecast: timesheets and planning analysis currency_id
Issue:
A traceback occurs when opening Timesheets and Planning Analysis.

Steps to reproduce:
1- Create a database with `project_timesheet_forecast_sale` installed.
2- Create a project and open the three-dot menu.
3- Click on `Timesheets and Planning Analysis`.

Cause:
The error occurs because currency_id field is not defined, as a
result:
https://github.com/odoo/odoo/blob/dff0a35413c0bcb106d0ab1086465f9200e25c5d/addons/web/static/src/views/pivot/pivot_renderer.js#L108-L116
currencyIds will cause a traceback as it is undefined.

This is due to planned_costs and effective_costs fields not being float but declared as widget="monetary" in the pivot view. Since there is no currency_field, the pivot renderer fails to resolve currency_id.

In stable in the pivot view we can remove monetary widget from fields.

opw-5176269

closes odoo/enterprise#97768

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>