[FIX] l10n_br_avatax: use custom views again
The actionable_errors widget was changed to ignore views if a
view_mode key is present in the action [1]. It caused the standard
product.product list view to pop up, which doesn't work in this
case. The custom views have the relevant Brazilian fields (SPED type
etc), can be edited inline, and have a "View" button allowing the user
to go to the product form view. The standard view has none of these.

[1] odoo/odoo@3bd833c73ea654d8558e3d9701ad41e4d75ddd6

opw-5154766

closes odoo/enterprise#98027

X-original-commit: 67fd490d619196f58b5a87d23c99cd1dfe949216
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Joren Van Onder (jov) <jov@odoo.com>