[FIX] account_accountant: Invalid regex in reconciliation rule creation
Currently, when a user uses the 'Set Account' button to reconcile a
statement line, the system may also create a reconciliation model
for fees for the given journal/account combination, based on the
previous statement line reference pattern.

To extract a patter from the references, the system computes the
normalised longest common substring. This will be used as regex
matching pattern of the newly created reconciliation rule.

However, the common substring may end up with '\\', which is not a valid
termination of a regex, leading to the error:
`Invalid operation. The regex is invalid.`

Steps to reproduce:

- Have an account '646000 Test Account'
- Create a Bank statement with label 'TEST REFERENCE BNK 2024'
- Set Account 646000
- Create a Bank statement with label 'TEST REFERENCE BNK.OTHER'
- Set Account 646000

Issue:

Error `Invalid operation. The regex is invalid.` will raise
blocking the action

opw-5086638

closes odoo/enterprise#97908

X-original-commit: 706e83e42d990ffef1debab5a08126603f817aa2
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>
Signed-off-by: Andrea Grazioso (agr) <agr@odoo.com>