[FIX] ai: force allow_end_message in tools of ai server actions
Purpose:
--------
The LLMApiService raises an error if no "text" message was sent by the
LLM during a request. This makes sense when using AI through a discuss
channel, to post a message in the chatter that the LLM did not provide an
answer.

However, for ai server actions, a "text" message is not always expected
(for example in ai_documents, some tool calls are done but no message is
expected from it).

For ai server actions, the "allow_end_message" is now forced to True so
that the loop in the llmApiService can break early because the llm can
notify that it is done.

Task-5144695

closes odoo/enterprise#96874

Signed-off-by: Stéphane Debauche (std) <std@odoo.com>