[IMP] {account,l10n_*}_reports: remove unaffected earnings from reports
This commit changes the way the Unaffected Earnings are displayed in
the Trial Balance and General Ledger. Instead of allocating unaffected
earnings to the equity_unaffected account when the report is rendered,
we now showcase the earnings to an additional line per company, with
no relation to an account.

The unaffected earnings account also needs to be removed from the account
returns audits.

Finally, we also adapt the trial balance reports and tests in localizations.

task-5153168

Part-of: odoo/enterprise#95062
Related: odoo/odoo#227754
Signed-off-by: William André (wan) <wan@odoo.com>