[FIX] iot: fix ingenico payments
Currently when paying with ingenco there is an error:

"Uncaught (in promise) TypeError: can't access property "payment_method",
line is undefined."

This is due to the fact that Ingenco sends requests to the database
every second even when the status of the driver didn't change since the
last payment.
This PR fixes the issue by only sending requests to the database if
a) THe driver status has changed
b) A payment is being processed

It also fixes the deprecation warning for isSet() replaced by is_set().

opw-5166439
opw-5181429
opw-5164612
opw-5170658

closes odoo/enterprise#98121

X-original-commit: eea573f962bcf7da244f59103fc83deb832f79ae
Signed-off-by: Louis Travaux (lotr) <lotr@odoo.com>
Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>