[FIX] pos_iot_six: undefined user_id opening session w/ different user
This commit fixes an undefined `user_id` on the pos session when trying to
start a payment with a Six terminal from another user than the one that started
the session.

Before this commit:

- open a pos session with Mitchell Admin,
- check the value of `pos.session.user_id` (it will be `2`),
- open another tab and connect as Marc Demo,
- check the value of `pos.session.user_id` again: it should be `6`, but instead
is `undefined`.

After this commit:

We use `pos.user.id` instead of `pos.session.user_id.id` to get the user ID.

opw-5055977

closes odoo/enterprise#98055

X-original-commit: 6b38fb947efb72d1414c9a44ae8b6ff0766ff3ed
Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>
Signed-off-by: Louis Travaux (lotr) <lotr@odoo.com>