[FIX] industry_fsm_sale: check project is accessible before checking project sharing
Before this commit, the portal user who has access to a fsm task and not
the project related could get an access error when he tries to access to
the form view of that task inside his portal. The reason is because a
check is made to know if the user has access to project sharing to
display the quotations button in the portal form view of the task.

This commit makes sure the project is in the parameter of the method in
which we did that check to make sure the user comes from a project
instead of `/my/tasks` route. Moreover, the quotation button will be
displayed once there is one quotation linked to the task. If there are
more than one then we will keep the current behavior, that is, only
display the button if the user has access to the project in edit mode.

closes odoo/enterprise#98122

X-original-commit: b84584fa8562a162d7de84a5fa88f6a57e2ca3b3
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>