[FIX] l10n_be_codabox: non applicable VAT
If a company is not subject to taxes, they may not have a VAT number.
In that case, the field can be left empty, such that the Company
Registry is used instead for the CodaBox connection. However, the case
where "/" (Non Applicable) was used as the VAT number was handled.

This commit now handles VAT="/" in the same way it handles no VAT at all
by using the company registry as a fallback. The commit also cleans up
how the company ID is used to avoid duplicated code by creating a
computed field.

opw-5164155

closes odoo/enterprise#98277

Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>