[FIX] pos_discount: global discount in split_bill_screen
Steps to reproduce:
- Order some products in a pos restaurant
- Add a global discount
- Split the order

Issue:
The global discount can be chosen to be splitted.

Fix:
The first fix is to apply the global discount each time we add a product.
The second fix is to hide the discount when splitting, and recalculate the
discount on each splitted order afterwards execpt when tranferring an order.

task-5189067

closes odoo/enterprise#98222

Related: odoo/odoo#233027
Signed-off-by: Stéphane Vanmeerhaeghe (stva) <stva@odoo.com>