[FIX] spreadsheet_edition: Avoid company currency rpc
The perfomance commit added in https://github.com/odoo/odoo/pull/151725
did not account for the version history action that does not inherit
from `AbstractSpreadsheetAction`.

this missing value trickled down to a bug only detectable in the
VersionHistory action and which is addressed in https://github.com/odoo/odoo/pull/232985

This commit adds the same performance to VersionHistory action.

Task-5187293

closes odoo/enterprise#98870

X-original-commit: b353bf02f0b00bb90204afe96e18548d0eb9a27f
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>