[FIX] l10n_cl_edi: accept numeric codes for currency in DTE imports
Steps to reproduce:
- Install "l10n_cl_edi" and switch to a Chilean company
- Go in Accounting > Vendor > Bills
- Import an XML with the line `<Moneda>013</Moneda>`
- Traceback

Cause:
`Moneda` can be the currency name code like `USD` but also a numeric code corresponding to the currency.

Solution:
Add a dictionary linking codes to the currency names and check the dictionary if `Moneda` is numeric.

Also add a fallback on `CLP` in case the value of `Moneda` fails to be translated to a currency. This avoids a traceback later when reading `vals['currency_id']`.

opw-5184950

closes odoo/enterprise#98514

X-original-commit: 8fb2454834d4371705e55331d5cfb2e20adf8760
Signed-off-by: Joren Van Onder (jov) <jov@odoo.com>
Signed-off-by: Mathieu Coutant (mcou) <mcou@odoo.com>