[FIX] *: adapt tests to mockFetch changes
This commit is associated to a change in the way mocked requests are
processed by Hoot; i.e. the content type is only inferred by the
"Content-Type" header, meaning that the '{ pure: true }' option becomes
useless in most cases.

This also means that some errors that should have been picked up by the
client are now properly wrapped in a JSON-RPC payload (instead of the
generic "connection lost" error) and are properly thrown, resulting in a
crash in some tests.

Affected tests have been adapted, either by fixing errors or removing
'pure: true' when not necessary.

closes odoo/enterprise#98075

Related: odoo/odoo#233055
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Signed-off-by: Julien Mougenot (jum) <jum@odoo.com>