[FIX] account_accountant: Recompute bank reconcile widget when records change
Issue:

when navigating through paginated records in the bank reconcilliation widget the functions `computeReconcileLineCountPerPartnerId` and `computeAvailableReconcileModels` were not re-executed. This caused reconciliation model buttons and partner counts to disappear on pages after the first.

steps to Reproduce:

1 - In a database with >80 records to reconcile and pre-existing reconciliation models:
2 - Go to Accounting -> N to reconcile button.
3 - In Kanban view expand a record to view reconciliation model buttons (displays correctly).
4 - Navigate to the next page.
 - From the second page onward, reconciliation model buttons and partner counts are not displayed.

fix:

on top of the `onWillStart` hook, which runs at the view is first rendered, I added `onRootLoaded` to ensure computations run whenever the dependencies change, updating the widget correctly on pagination.

opw-5083877

closes odoo/enterprise#98916

X-original-commit: 5358bee078b4bb15ea5ce82e21771c24b8cdaa6d
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>