[FIX] pos_l10n_se: fix blackbox order validation
Before this commit, the Swedish blackbox was completely broken in 18.0
because the `push_single_order` function is no longer used.

To fix this, the same approach as `pos_blackbox_be` was used, where
the `preSyncAllOrders` function was overridden instead.
In addition, the old receipt reprinting logic was no longer working, and
so it now uses the `nb_print` field to keep track of the number of
reprints.

Finally, we also forward port a fix from 17.0 (ddbc1fc) to correct the
format of the organisation number, which for some reason was not forward
ported originally.

task-5077448

closes odoo/enterprise#98437

X-original-commit: f936f1bfd042403da162e9b8b00ec4feb65cd6ad
Signed-off-by: Louis Travaux (lotr) <lotr@odoo.com>
Signed-off-by: Max Whale (mawh) <mawh@odoo.com>