[IMP] account_reports: show account_return total balance if different
When this https://github.com/odoo/enterprise/pull/91886 got merged,
It removed the accumulation of the balance in the return amount. Showing only the amount of the current period.
Additionally, the total amount was shown if the amount was different. But, it was only being shown if it was less than the current period amount.

The fix ensures it is always shown if the amount is different.

task-5172274

closes odoo/enterprise#98150

Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>