[FIX] stock_barcode_mrp: consume smls
Steps to reproduce:
1- Create MO
2- Change the `qty_producing`

Issue:
`stock.move.lines` are not consumed. Because `qty_producing` is not a
computed field therefore it has no inverse. It updates the consumption
with an on change method and in Barcode we don't have `move_raw_ids` in
the xml, so its not stored or saved.

To fix the problem, `set_qty_producing` was called manually to keep the
barcode's design clean.

closes odoo/enterprise#98184

Task: 5111357
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>